+++ deepxube/tests/test_env.py
@@ -6,6 +6,7 @@ from torch.multiprocessing import Queue, get_context
 
 from deepxube.environments.environment_abstract import Environment, State
 from deepxube.nnet import nnet_utils
+from deepxube.utils.misc_utils import flatten
 import numpy as np
 from numpy.typing import NDArray
 
@@ -33,33 +34,39 @@ def test_env(env: Environment[Any, Any], num_states: int, step_max: int):
 
     # get data
     start_time = time.time()
-    states, goals = env.get_start_goal_pairs(list(np.random.randint(step_max, size=num_states)))
+    states, goals = env.get_start_goal_pairs(list(np.random.randint(step_max + 1, size=num_states)))
 
     elapsed_time = time.time() - start_time
     states_per_sec = len(states) / elapsed_time
     print("Generated %i start/goal states in %s seconds (%.2f/second)" % (len(states), elapsed_time, states_per_sec))
 
-    # expand
+    # is_solved
     start_time = time.time()
-    env.expand(states)
+    is_solved_l: List[bool] = env.is_solved(states, goals)
+    per_solved: float = 100.0 * float(np.mean(is_solved_l))
     elapsed_time = time.time() - start_time
     states_per_sec = len(states) / elapsed_time
-    print("Expanded %i states in %s seconds (%.2f/second)" % (len(states), elapsed_time, states_per_sec))
+    print(f"Computed is_solved for {len(states)} states ({per_solved:.2f}% solved) in {elapsed_time} seconds "
+          f"({states_per_sec:.2f}/second)")
+
+    # expand
+    start_time = time.time()
+    states_exp, tcs = env.expand(states)
+    ave_next_states: float = float(np.mean([len(x) for x in states_exp]))
+    ave_tc: float = float(np.mean(flatten(tcs)[0]))
+
+    elapsed_time = time.time() - start_time
+    states_per_sec = len(states) / elapsed_time
+    print(f"Expanded %i states, mean #next/tc: ({ave_next_states:.2f}/{ave_tc:.2f}), "
+          f"in %s seconds (%.2f/second)" % (len(states), elapsed_time, states_per_sec))
 
     # nnet format
     start_time = time.time()
-    states_nnet: List[NDArray[Any]] = env.states_to_nnet_input(states)
+    states_goals_nnet: List[NDArray[Any]] = env.states_goals_to_nnet_input(states, goals)
     elapsed_time = time.time() - start_time
     states_per_sec = len(states) / elapsed_time
-    print("Converted %i states to nnet format in "
+    print("Converted %i states and goals to nnet format in "
           "%s seconds (%.2f/second)" % (len(states), elapsed_time, states_per_sec))
-
-    start_time = time.time()
-    goals_nnet = env.goals_to_nnet_input(goals)
-    elapsed_time = time.time() - start_time
-    states_per_sec = len(goals) / elapsed_time
-    print("Converted %i goals to nnet format in "
-          "%s seconds (%.2f/second)" % (len(goals), elapsed_time, states_per_sec))
 
     # get heuristic fn
     on_gpu: bool
@@ -75,11 +82,11 @@ def test_env(env: Environment[Any, Any], num_states: int, step_max: int):
     # initialize nnet
     print("")
     heuristic_fn = nnet_utils.get_heuristic_fn(nnet, device, env)
-    heuristic_fn(states_nnet, goals_nnet)
+    heuristic_fn(states, goals)
 
     # nnet heuristic
     start_time = time.time()
-    heuristic_fn(states_nnet, goals_nnet)
+    heuristic_fn(states, goals)
 
     nnet_time = time.time() - start_time
     states_per_sec = len(states) / nnet_time
@@ -96,7 +103,7 @@ def test_env(env: Environment[Any, Any], num_states: int, step_max: int):
     proc.start()
     print("Process start time: %.2f" % (time.time() - start_time))
 
-    queue1.put((states_nnet, goals_nnet))
+    queue1.put((states_goals_nnet,))
     queue2.get()
 
     start_time = time.time()
@@ -108,7 +115,7 @@ def test_env(env: Environment[Any, Any], num_states: int, step_max: int):
     print("States/goals get time: %.2f" % (time.time() - start_time))
 
     start_time = time.time()
-    queue1.put((states_nnet, goals_nnet))
+    queue1.put((states_goals_nnet,))
     print("State/goals nnet send time: %s" % (time.time() - start_time))
 
     start_time = time.time()