+++ deepxube/utils/timing_utils.py
@@ -97,11 +97,11 @@ class Times:
         return time_tot
 
     def get_time_str(self, prefix: str = "") -> str:
-        time_str_l: List[str] = ["%s: %.2f" % (key, val) for key, val in self.times.items()]
-        sub_time_str_l: List[str] = ["->%s: %.2f" % (key, sub_time.get_total_time())
+        time_str_l: List[str] = ["%s: %f" % (key, val) for key, val in self.times.items()]
+        sub_time_str_l: List[str] = ["->%s: %f" % (key, sub_time.get_total_time())
                                      for key, sub_time in self.sub_times.items()]
 
-        time_str: str = ", ".join(time_str_l + sub_time_str_l + ["Tot: %.2f" % self.get_total_time()])
+        time_str: str = ", ".join(time_str_l + sub_time_str_l + ["Tot: %f" % self.get_total_time()])
 
         prefix_new: str = f"\t{prefix}"
         for key, sub_time in self.sub_times.items():